feat(dashboard-shared): separate dashboard detail from dashboard detail page
